+++ src/test/java/com/nubiform/order/controller/SignUpInvalidTest.java
@@ -15,6 +15,7 @@ import static com.nubiform.order.controller.AuthController.API_V1_AUTH_URI;
 import static com.nubiform.order.controller.AuthController.SIGN_UP;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SpringBootTest
@@ -52,7 +53,10 @@ class SignUpInvalidTest {
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(objectMapper.writeValueAsString(signUpRequest)))
                 .andDo(print())
-                .andExpect(status().isBadRequest());
+                .andExpect(status().isBadRequest())
+                .andExpect(jsonPath("$.code").value(1001))
+                .andExpect(jsonPath("$.description").value("invalid parameter"))
+                .andExpect(jsonPath("$.details[0].field").value("username"));
     }
 
     @Test
@@ -63,7 +67,10 @@ class SignUpInvalidTest {
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(objectMapper.writeValueAsString(signUpRequest)))
                 .andDo(print())
-                .andExpect(status().isBadRequest());
+                .andExpect(status().isBadRequest())
+                .andExpect(jsonPath("$.code").value(1001))
+                .andExpect(jsonPath("$.description").value("invalid parameter"))
+                .andExpect(jsonPath("$.details[0].field").value("nickname"));
     }
 
     @Test
@@ -74,7 +81,10 @@ class SignUpInvalidTest {
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(objectMapper.writeValueAsString(signUpRequest)))
                 .andDo(print())
-                .andExpect(status().isBadRequest());
+                .andExpect(status().isBadRequest())
+                .andExpect(jsonPath("$.code").value(1001))
+                .andExpect(jsonPath("$.description").value("invalid parameter"))
+                .andExpect(jsonPath("$.details[0].field").value("password"));
     }
 
     @Test
@@ -85,7 +95,10 @@ class SignUpInvalidTest {
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(objectMapper.writeValueAsString(signUpRequest)))
                 .andDo(print())
-                .andExpect(status().isBadRequest());
+                .andExpect(status().isBadRequest())
+                .andExpect(jsonPath("$.code").value(1001))
+                .andExpect(jsonPath("$.description").value("invalid parameter"))
+                .andExpect(jsonPath("$.details[0].field").value("phone"));
     }
 
     @Test
@@ -96,6 +109,27 @@ class SignUpInvalidTest {
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(objectMapper.writeValueAsString(signUpRequest)))
                 .andDo(print())
-                .andExpect(status().isBadRequest());
+                .andExpect(status().isBadRequest())
+                .andExpect(jsonPath("$.code").value(1001))
+                .andExpect(jsonPath("$.description").value("invalid parameter"))
+                .andExpect(jsonPath("$.details[0].field").value("email"));
+    }
+
+    @Test
+    void invalidAllTest() throws Exception {
+        signUpRequest.setUsername("12345!@#$%");
+        signUpRequest.setNickname("12345!@#$%");
+        signUpRequest.setPassword("password");
+        signUpRequest.setPhone("phone");
+        signUpRequest.setEmail("email");
+
+        mockMvc.perform(post(API_V1_AUTH_URI + SIGN_UP)
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(objectMapper.writeValueAsString(signUpRequest)))
+                .andDo(print())
+                .andExpect(status().isBadRequest())
+                .andExpect(jsonPath("$.code").value(1001))
+                .andExpect(jsonPath("$.description").value("invalid parameter"))
+                .andExpect(jsonPath("$.details.length()").value(5));
     }
 }
